create NewQuizDialog component
